requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Database
supabase>=2.0.0

# Payments
stripe>=7.0.0

# HTTP Client
httpx>=0.26.0

# Settings Management
pydantic-settings>=2.0.0

# NLP & Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Topic Modeling & Text Processing
scikit-learn>=1.3.0
gensim>=4.3.0

# Utilities
python-dateutil>=2.8.2
